feat: removes senate email campaign banner and CTA
